refactor table extractor to be own class that returns list of tables w/ cacheing; instantiates using pdf-to-xml-parser and xml-to-tables-parser classes
