+++ corvid/table_extraction/pdf_to_xml_parser.py
@@ -1,6 +1,7 @@
 """
 
-Classes that calls local PDF -> XML parser via subprocess
+Classes that parse local PDF -> XML by subprocess to call external parsing tool.
+Typical usage is within TableExtractor that utilizes external PDF parsing tools.
 
 """
 
@@ -8,19 +9,19 @@ import os
 import subprocess
 
 
-class PDFParserException(Exception):
+class PDFToXMLParserException(Exception):
     pass
 
 
-class TetmlPDFParserException(PDFParserException):
+class TetmlPDFToXMLParserException(PDFToXMLParserException):
     pass
 
 
-class OmnipagePDFParserException(PDFParserException):
+class OmnipagePDFToXMLParserException(PDFToXMLParserException):
     pass
 
 
-class PDFParser(object):
+class PDFToXMLParser(object):
     def __init__(self, target_dir: str):
         if not os.path.exists(target_dir):
             raise FileNotFoundError('Target directory {} doesnt exist'
@@ -28,35 +29,37 @@ class PDFParser(object):
         self.target_dir = target_dir
 
     def parse(self, paper_id: str, source_pdf_path: str) -> str:
-        """Primary method for parsing a Paper's PDF given its id.
-        Returns the local path of the XML output of the parsing.
-        Raises exception unless user implements `_parse()`
+        """Parses a Paper's PDF to XML given its paper_id.
+        Returns the local path of the resulting XML output.
+        Raises exception unless user implements `_parse()`.
         """
         target_path = self.get_target_path(paper_id)
         self._parse(paper_id, source_pdf_path, target_path)
         return target_path
 
-    def _parse(self, paper_id: str, source_pdf_path: str, target_path: str):
-        """User should implement this.  Recommended to catch Exceptions
-        and instead raise custom Exceptions based on PDFParserException."""
+    def _parse(self, paper_id: str, source_pdf_path: str, target_xml_path: str):
+        """User should implement this.
+        Recommended to catch Exceptions and then raise custom Exceptions
+        based on PDFParserException.
+        """
         raise NotImplementedError
 
     def get_target_path(self, paper_id: str) -> str:
-        """Each PDF parser is responsible for constructing output filenames
-        given the paper_id and self.target_dir.  This includes any special
-        nesting directories (e.g. multiple files output given paper_id)"""
+        """Each PDF to XML parser is responsible for constructing output
+        filenames given the paper_id and self.target_dir.  This includes any
+        special nesting directories (e.g. multiple files output given paper_id)
+        """
         raise NotImplementedError
 
 
-class TetmlPDFParser(PDFParser):
+class TetmlPDFToXMLParser(PDFToXMLParser):
     def __init__(self, tet_bin_path: str, target_dir: str):
         if not os.path.exists(tet_bin_path):
-            raise TetmlPDFParserException(
-                '{} doesnt exist'.format(tet_bin_path))
+            raise FileNotFoundError('{} doesnt exist'.format(tet_bin_path))
         self.tet_bin_path = tet_bin_path
-        super(TetmlPDFParser, self).__init__(target_dir)
+        super(TetmlPDFToXMLParser, self).__init__(target_dir)
 
-    def _parse(self, paper_id: str, source_pdf_path: str, target_path: str):
+    def _parse(self, paper_id: str, source_pdf_path: str, target_xml_path: str):
         try:
             cmd = '{tet} --tetml wordplus --targetdir {targetdir} --pageopt {pageopt} --docopt checkglyphlists {pdf}' \
                 .format(tet=self.tet_bin_path,
@@ -65,35 +68,34 @@ class TetmlPDFParser(PDFParser):
                         pdf=source_pdf_path)
             subprocess.run(cmd, shell=True, check=True)
         except subprocess.CalledProcessError as e:
-            if os.path.exists(target_path):
-                os.remove(target_path)
+            if os.path.exists(target_xml_path):
+                os.remove(target_xml_path)
             print(e)
-            raise TetmlPDFParserException('TET failed to parse {}'
-                                          .format(source_pdf_path))
+            raise TetmlPDFToXMLParserException('TET failed to parse {}'
+                                               .format(source_pdf_path))
 
     def get_target_path(self, paper_id: str) -> str:
         return '{}.tetml'.format(os.path.join(self.target_dir, paper_id))
 
 
-class OmnipagePDFParser(PDFParser):
+class OmnipagePDFToXMLParser(PDFToXMLParser):
     def __init__(self, omnipage_bin_path: str, target_dir: str):
         if not os.path.exists(omnipage_bin_path):
-            raise OmnipagePDFParserException('{} doesnt exist'
-                                             .format(omnipage_bin_path))
+            raise FileNotFoundError('{} doesnt exist'.format(omnipage_bin_path))
         self.omnipage_bin_path = omnipage_bin_path
-        super(OmnipagePDFParser, self).__init__(target_dir)
+        super(OmnipagePDFToXMLParser, self).__init__(target_dir)
 
-    def _parse(self, paper_id: str, source_pdf_path: str, target_path: str):
+    def _parse(self, paper_id: str, source_pdf_path: str, target_xml_path: str):
         try:
             cmd = '{bin} -i {input_path} -o {output_path}' \
                 .format(bin=self.omnipage_bin_path,
                         input_path=source_pdf_path,
-                        output_path=target_path)
+                        output_path=target_xml_path)
             subprocess.run(cmd, shell=True, check=True)
         except subprocess.CalledProcessError as e:
             print(e)
-            raise OmnipagePDFParserException('OmniPage failed to parse {}'
-                                             .format(source_pdf_path))
+            raise OmnipagePDFToXMLParserException('OmniPage failed to parse {}'
+                                                  .format(source_pdf_path))
 
     def get_target_path(self, paper_id: str) -> str:
         return '{}.xml'.format(os.path.join(self.target_dir, paper_id))
